requirements.txt
openai
pyyaml
pandas
scikit-learn
transformers
numpy